Move UserModel to a models.py file
