+++ server/infrastructure/auth/repositories.py
@@ -1,47 +1,17 @@
-import uuid
-from typing import TYPE_CHECKING, Any, Optional
+from typing import Any, Optional
 
-from sqlalchemy import CHAR, Column, Enum, ForeignKey, String, delete, select
-from sqlalchemy.dialects.postgresql import UUID
+from sqlalchemy import delete, select
 from sqlalchemy.exc import NoResultFound
 from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy.orm import relationship
 
-from server.application.auth.passwords import API_TOKEN_LENGTH
-from server.domain.auth.entities import User, UserRole
+from server.domain.auth.entities import User
 from server.domain.auth.exceptions import UserDoesNotExist
 from server.domain.auth.repositories import UserRepository
 from server.domain.common.types import ID
-from server.domain.organizations.types import Siret
 
-from ..database import Base, Database
-
-if TYPE_CHECKING:
-    from ..organizations.models import OrganizationModel
-
-
-class UserModel(Base):
-    __tablename__ = "user"
-
-    id: uuid.UUID = Column(UUID(as_uuid=True), primary_key=True)
-    organization_siret: Siret = Column(
-        CHAR(14),
-        ForeignKey("organization.siret"),
-        nullable=False,
-    )
-    organization: "OrganizationModel" = relationship(
-        "OrganizationModel",
-        back_populates="users",
-    )
-    email = Column(String, nullable=False, unique=True, index=True)
-    password_hash = Column(String, nullable=False)
-    role = Column(Enum(UserRole, name="user_role_enum"), nullable=False)
-    api_token = Column(String(API_TOKEN_LENGTH), nullable=False)
-
-
-def update_instance(instance: UserModel, entity: User) -> None:
-    for field in set(User.__fields__) - {"id"}:
-        setattr(instance, field, getattr(entity, field))
+from ..database import Database
+from .models import UserModel
+from .transformers import update_instance
 
 
 class SqlUserRepository(UserRepository):
